SendMessageInterceptorImpl repos refactor
