+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/plugin/factory/StreamOfflinePluginFactory.kt
@@ -109,7 +109,9 @@ public class StreamOfflinePluginFactory(
             context = appContext,
             logic = logic,
             globalState = globalState,
-            repos = repos,
+            channelRepository = repos,
+            messageRepository = repos,
+            attachmentRepository = repos,
             scope = scope,
             networkType = config.uploadAttachmentsNetworkType
         )
